requirements.txt
streamlit
python-binance
pandas
numpy
matplotlib
praw
yfinance
textblob
requests
